requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
Django==2.2
django-cors-headers==3.0.2
django-filter==2.2.0
djangorestframework==3.10.1
jwcrypto==0.6.0
Markdown==3.1.1
Pillow==6.2.1
pkg-resources==0.0.0
psycopg2==2.8.3
pycparser==2.19
python-dateutil==2.8.0
python-http-client==3.2.1
pytz==2019.1
sendgrid==6.1.0
six==1.12.0
sqlparse==0.3.0
